firejail: Add package at version 0.9.70
I've included two of my own patches. One swaps a hard dependency on
gawk for generic awk, which seems to work fine with Busybox's awk.
The other fixes POSIX compliance in the mkman.sh script, which
depends on non-POSIX uses of sed and date. sed usage is easy enough
to fix, but POSIX date does barely anything compared to GNU date
and the easiest solution I've found in the past is just to use
Python instead.