firejail: Remove build-time dependency on Python
I realised the date invocations in mkman.sh were literally just
to generate the date at the bottom of the manpage, which is a
ridiculous thing to require a whole dependency on Python. I've
updated the POSIX compliance patch so that we use the date utility
and just get the current month and year, ignoring the SOURCE_DATE_EPOCH
variable.