Purge fwupd and its dependencies
I just can't get fwupd working on my system. It compiles and runs,
but in addition to requiring forks of a bunch of community packages,
it also requires dbus. Then I couldn't get it to detect anything
on my system without installing udisks, which would in turn require
polkit. The whole thing is just too much. I've done this in one commit
so it can be easily reverted if someone wants to fork/maintain it.

The removed packages are:
- fwupd-efi
- fwupd
- tpm2-tss
- freetype-harfbuzz
- gcab
- gnu-efi
- gobject-introspection
- json-glib
- libftdi1
- libgudev
- pango
- protobuf-c
- pycairo
- pygobject